+++ src/main/java/com/imsweb/validation/entities/SimpleMapValidatable.java
@@ -3,7 +3,7 @@
  */
 package com.imsweb.validation.entities;
 
-import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -12,8 +12,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.SortedSet;
-
-import org.apache.commons.beanutils.PropertyUtils;
 
 import com.imsweb.validation.ValidatorServices;
 import com.imsweb.validation.internal.ExtraPropertyEntityHandlerDto;
@@ -242,20 +240,29 @@ public class SimpleMapValidatable implements Validatable {
                     int colPos = javaPath == null ? -1 : javaPath.lastIndexOf(".");
                     if (javaPath != null && colPos >= 0) {
                         String collectionName = javaPath.substring(colPos + 1);
-                        try {
-                            Object collection = PropertyUtils.getProperty(_current, collectionName);
-                            if (collection != null) {
-                                int index = -1;
-                                if (collection instanceof List)
-                                    index = ((List<Object>)collection).indexOf(extra.getEntity());
-                                else if (collection instanceof SortedSet)
-                                    index = new ArrayList<>((SortedSet<Object>)collection).indexOf(extra.getEntity());
-                                if (index != -1)
-                                    _propertiesWithError.add(_prefix + "." + collectionName + "[" + index + "]." + prop.substring(pos + 1));
+
+                        Object collection;
+                        if (_current instanceof Map<?, ?>)
+                            collection = ((Map<?, ?>)_current).get(collectionName);
+                        else {
+                            try {
+                                Field field = _current.getClass().getDeclaredField(collectionName);
+                                field.setAccessible(true);
+                                collection = field.get(_current);
+                            }
+                            catch (IllegalAccessException | NoSuchFieldException e) {
+                                throw new IllegalAccessException("Unable to access " + collectionName + " on object of type + " + _current.getClass().getName());
                             }
                         }
-                        catch (IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
-                            throw new IllegalAccessException("Unable to access " + collectionName + " on object of type + " + _current.getClass().getName());
+
+                        if (collection != null) {
+                            int index = -1;
+                            if (collection instanceof List)
+                                index = ((List<Object>)collection).indexOf(extra.getEntity());
+                            else if (collection instanceof SortedSet)
+                                index = new ArrayList<>((SortedSet<Object>)collection).indexOf(extra.getEntity());
+                            if (index != -1)
+                                _propertiesWithError.add(_prefix + "." + collectionName + "[" + index + "]." + prop.substring(pos + 1));
                         }
                     }
                 }